requirements.txt
ansible-compat==2.2.7
ansible-core==2.14.1
ansible-lint==6.11.0
attrs==22.2.0
black==22.12.0
bracex==2.3.post1
cffi==1.15.1
click==8.1.3
commonmark==0.9.1
cryptography==39.0.0
exceptiongroup==1.1.0
filelock==3.9.0
iniconfig==2.0.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
pycparser==2.21
Pygments==2.14.0
pyrsistent==0.19.3
pytest==7.2.1
pytest-testinfra==7.0.0
PyYAML==6.0
resolvelib==0.8.1
rich==13.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
subprocess-tee==0.4.1
tomli==2.0.1
wcmatch==8.4.1
yamllint==1.29.0
